# test(tm2/pkg/cmap): add benchmarks to show true impact of contention