@jefft0 2025-01-22
@odeke-em , can you resolve the CI check error for "TM2 / Run TM2 suite / Go Lint" ?